+++ src/Models/ShippingFileMovements.php
@@ -1,0 +1,29 @@
+<?php
+/**
+ * This Software is part of aryelgois/bank-interchange and is provided "as is".
+ *
+ * @see LICENSE
+ */
+
+namespace aryelgois\BankInterchange\Models;
+
+use aryelgois\Medools;
+
+/**
+ * Creates a relation of Titles to Shipping Files
+ *
+ * @author Aryel Mota Góis
+ * @license MIT
+ * @link https://www.github.com/aryelgois/bank-interchange
+ */
+class ShippingFileMovements extends Medools\Model
+{
+    const TABLE = 'shipping_file_movements';
+
+    const COLUMNS = [
+        'id',
+        'cnab',
+        'code',
+        'name',
+    ];
+}